requirements.txt
faiss-cpu>=1.12.0
numpy>=1.25.0
volcengine-python-sdk[ark]>=4.0.31
streamlit>=1.28.0
Pillow>=10.0.0

